Refactor run/build startup

+++ misc/config_base.sh
@@ -5,7 +5,7 @@ DEFAULT_CONF=${DAQ_CONF:-misc/system_base.conf}
 
 run_mode=$(cat misc/RELEASE_VERSION)
 
-if [ ! -f "$LOCAL_SYSTEM" ]; then
+if [ ! -f "$LOCAL_SYSTEM" -a -z "$DONT_MAKE_LOCAL" ]; then
     echo No $LOCAL_SYSTEM found, copying defaults from $DEFAULT_CONF...
     mkdir -p local
     cp $DEFAULT_CONF $LOCAL_SYSTEM
